backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "source-lens-api"
version = "0.1.0"
description = "Backend scaffold for the Source Lens API."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "beautifulsoup4>=4.13,<5",
  "fastapi>=0.115,<1",
  "httpx>=0.28,<1",
  "pydantic-settings>=2.8,<3",
  "pypdf>=5.4,<6",
  "qdrant-client>=1.11,<2",
  "uvicorn>=0.34,<1",
]

[project.optional-dependencies]
dev = [
  "pyright>=1.1.390,<2",
  "pytest>=8.3,<9",
  "ruff>=0.11,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/source_lens_api"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pyright]
include = ["src", "tests"]
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"
